+++ mne_icalabel/utils/_checks.py
